refactor(telegram): extract s5 ingest/media clusters into mixin modules (epic #78791)
