+++ plugins/platforms/telegram/reactions_mixin.py
@@ -1,0 +1,115 @@
+"""Message reactions (processing lifecycle) methods for ``TelegramAdapter``.
+
+Extracted from ``plugins/platforms/telegram/adapter.py`` as part of the
+god-file decomposition campaign, following the same mechanical mixin lift
+that produced ``gateway/authz_mixin.py`` and the Telegram authorization
+mixin (PR #75742). This mixin holds the c13 cluster: the 👀/👍/👎 processing-lifecycle reactions driven by the ``on_processing_start`` / ``on_processing_complete`` base hooks, gated by ``TELEGRAM_REACTIONS``.
+
+Behavior-neutral: every method is lifted verbatim from ``TelegramAdapter``.
+Class attributes (none) stay on ``TelegramAdapter`` and
+resolve via ``self.*`` / ``cls.*`` through the MRO, exactly as before the
+lift, and ``ReactionsMixin`` precedes ``BasePlatformAdapter`` in the bases.
+
+``logger`` is bound by explicit name so records emitted from these methods
+keep the logger name ``"plugins.platforms.telegram.adapter"``. ``Message``
+is imported under the same ``ImportError`` guard the adapter uses, falling
+back to ``Any``; like the adapter, this module does not enable postponed
+annotation evaluation.
+"""
+
+import logging
+import os
+from typing import Any
+
+from gateway.platforms.base import (
+    MessageEvent,
+    ProcessingOutcome,
+)
+from plugins.platforms.telegram.telegram_ids import normalize_telegram_chat_id
+
+logger = logging.getLogger("plugins.platforms.telegram.adapter")
+
+class ReactionsMixin:
+    """Message reactions (processing lifecycle) cluster lifted verbatim from ``TelegramAdapter``."""
+
+    # ── Message reactions (processing lifecycle) ──────────────────────────
+
+    def _reactions_enabled(self) -> bool:
+        """Check if message reactions are enabled via config/env."""
+        return os.getenv("TELEGRAM_REACTIONS", "false").lower() not in {"false", "0", "no"}
+
+    async def _set_reaction(self, chat_id: str, message_id: str, emoji: str) -> bool:
+        """Set a single emoji reaction on a Telegram message."""
+        if not self._bot:
+            return False
+        try:
+            await self._bot.set_message_reaction(
+                chat_id=normalize_telegram_chat_id(chat_id),
+                message_id=int(message_id),
+                reaction=emoji,
+            )
+            return True
+        except Exception as e:
+            logger.debug("[%s] set_message_reaction failed (%s): %s", self.name, emoji, _redact_telegram_error_text(e))
+            return False
+
+    async def _clear_reactions(self, chat_id: str, message_id: str) -> bool:
+        """Clear all reactions from a Telegram message.
+
+        Calling ``set_message_reaction`` with ``reaction=None`` (or an empty
+        sequence) is the documented Bot API way to remove all bot-set
+        reactions on a message — equivalent to Bot API 10.0's
+        ``deleteMessageReaction`` but supported in PTB 22.6 already.
+        """
+        if not self._bot:
+            return False
+        try:
+            await self._bot.set_message_reaction(
+                chat_id=normalize_telegram_chat_id(chat_id),
+                message_id=int(message_id),
+                reaction=None,
+            )
+            return True
+        except Exception as e:
+            logger.debug("[%s] clear reactions failed: %s", self.name, _redact_telegram_error_text(e))
+            return False
+
+    async def on_processing_start(self, event: MessageEvent) -> None:
+        """Add an in-progress reaction when message processing begins."""
+        if not self._reactions_enabled():
+            return
+        chat_id = getattr(event.source, "chat_id", None)
+        message_id = getattr(event, "message_id", None)
+        if chat_id and message_id:
+            await self._set_reaction(chat_id, message_id, "\U0001f440")
+
+    async def on_processing_complete(self, event: MessageEvent, outcome: ProcessingOutcome) -> None:
+        """Swap the in-progress reaction for a final success/failure reaction.
+
+        Unlike Discord (additive reactions), Telegram's set_message_reaction
+        replaces all existing reactions in one call — no remove step needed.
+
+        On CANCELLED outcomes (e.g. the user runs ``/stop``, or a session is
+        interrupted mid-flight), we explicitly clear the 👀 in-progress
+        reaction so it doesn't linger on the user's message indefinitely.
+        Without this clear, the only way to remove the 👀 was to wait for
+        another agent run to swap it to 👍/👎 — which never happens if the
+        cancellation was the last activity in the chat.
+        """
+        if not self._reactions_enabled():
+            return
+        chat_id = getattr(event.source, "chat_id", None)
+        message_id = getattr(event, "message_id", None)
+        if not (chat_id and message_id):
+            return
+        if outcome == ProcessingOutcome.CANCELLED:
+            await self._clear_reactions(chat_id, message_id)
+        else:
+            await self._set_reaction(
+                chat_id,
+                message_id,
+                "\U0001f44d" if outcome == ProcessingOutcome.SUCCESS else "\U0001f44e",
+            )
+
+
+from plugins.platforms.telegram.adapter import _redact_telegram_error_text
